# api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth.me": {
      "get": {
        "operationId": "UsersController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current user profile"
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Get current user profile",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth.updateProfile": {
      "post": {
        "operationId": "UsersController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated user profile"
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Update current user profile",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth.changePassword": {
      "post": {
        "operationId": "UsersController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed successfully"
          },
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Change current user password",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/smtp.status": {
      "get": {
        "operationId": "SmtpController_status",
        "parameters": [
          {
            "name": "workspace_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SMTP status"
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Check if SMTP is available",
        "tags": [
          "smtp"
        ]
      }
    },
...
    "/api/smtp.update": {
      "post": {
        "operationId": "SmtpController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSmtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated SMTP settings"
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SmtpSettingsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Update SMTP settings (owner only)",
        "tags": [
          "smtp"
        ]
      }
    },
    "/api/smtp.delete": {
      "post": {
        "operationId": "SmtpController_delete",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteSmtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "SMTP settings removed"
          },
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Remove workspace SMTP (fall back to global)",
        "tags": [
          "smtp"
        ]
      }
    },
    "/api/smtp.test": {
      "post": {
        "operationId": "SmtpController_test",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestSmtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Test email sent"
          },
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Send test email",
        "tags": [
          "smtp"
        ]
      }
    },
    "/api/workspaces.list": {
      "get": {
        "operationId": "WorkspacesController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "List all workspaces",
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/workspaces.get": {
      "get": {
        "operationId": "WorkspacesController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Get workspace by ID",
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/workspaces.create": {
      "post": {
        "operationId": "WorkspacesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspaceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Create a new workspace",
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/workspaces.update": {
      "post": {
        "operationId": "WorkspacesController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkspaceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Update an existing workspace",
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/workspaces.delete": {
      "post": {
        "operationId": "WorkspacesController_delete",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  }
                },
                "required": [
                  "id"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Delete a workspace",
        "tags": [
          "workspaces"
        ]
      }
    },
...
    "/api/tools.websiteMeta": {
      "post": {
        "operationId": "ToolsController_websiteMeta",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebsiteMetaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Website metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebsiteMetaResponse"
                }
              }
            }
          },
          "400": {
            "description": "Failed to fetch metadata"
          }
        },
        "security": [
          {}
        ],
        "summary": "Fetch website title and logo",
        "tags": [
          "tools"
        ]
      }
    },
    "/api/tools.favicon": {
      "get": {
        "operationId": "ToolsController_favicon",
        "parameters": [
          {
            "name": "url",
            "required": true,
            "in": "query",
            "description": "Website URL to fetch favicon for",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Favicon image binary"
          }
        },
        "security": [
          {}
        ],
        "summary": "Fetch and proxy website favicon",
        "tags": [
          "tools"
        ]
      }
    },
...
  },
  "info": {
    "title": "Staminads API",
    "description": "Web analytics platform for tracking TimeScore metrics",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "jwt-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      },
      "demo-secret": {
        "type": "apiKey",
        "in": "query",
        "name": "secret",
        "description": "Demo secret for demo endpoints"
      }
    },
    "schemas": {
...
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          }
        }
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string"
          },
          "newPassword": {
            "type": "string",
            "minLength": 8,
            "maxLength": 72
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
      "UpdateSmtpDto": {
        "type": "object",
        "properties": {
          "workspace_id": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "host": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          },
          "port": {
            "type": "number",
            "minimum": 1,
            "maximum": 65535
          },
          "username": {
            "type": "string",
            "maxLength": 255
          },
          "password": {
            "type": "string",
            "maxLength": 255
          },
          "from_name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "from_email": {
            "type": "string",
            "format": "email"
          }
        },
        "required": [
          "workspace_id",
          "enabled",
          "host",
          "port",
          "from_name",
          "from_email"
        ]
      },
      "SmtpSettingsDto": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "host": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          },
          "port": {
            "type": "number",
            "minimum": 1,
            "maximum": 65535
          },
          "username": {
            "type": "string",
            "maxLength": 255
          },
          "password": {
            "type": "string",
            "maxLength": 255
          },
          "from_name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "from_email": {
            "type": "string",
            "format": "email"
          }
        },
        "required": [
          "enabled",
          "host",
          "port",
          "from_name",
          "from_email"
        ]
      },
      "DeleteSmtpDto": {
        "type": "object",
        "properties": {
          "workspace_id": {
            "type": "string"
          }
        },
        "required": [
          "workspace_id"
        ]
      },
      "TestSmtpDto": {
        "type": "object",
        "properties": {
          "workspace_id": {
            "type": "string"
          },
          "to_email": {
            "type": "string",
            "format": "email"
          }
        },
        "required": [
          "workspace_id",
          "to_email"
        ]
      },
      "AnnotationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "date": {
            "type": "string"
          },
          "timezone": {
            "type": "string"
          },
          "title": {
            "type": "string",
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "maxLength": 500
          },
          "color": {
            "type": "string",
            "pattern": "/^#[0-9A-Fa-f]{6}$/"
          }
        },
        "required": [
          "id",
          "date",
          "timezone",
          "title"
        ]
      },
      "CreateWorkspaceSettingsDto": {
        "type": "object",
        "properties": {
          "timescore_reference": {
            "type": "number",
            "minimum": 1
          },
          "bounce_threshold": {
            "type": "number",
            "minimum": 1
          },
          "geo_enabled": {
            "type": "boolean"
          },
          "geo_store_city": {
            "type": "boolean"
          },
          "geo_store_region": {
            "type": "boolean"
          },
          "geo_coordinates_precision": {
            "type": "number",
            "minimum": 0,
            "maximum": 2
          },
          "annotations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AnnotationDto"
            }
          }
        }
      },
      "CreateWorkspaceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50,
            "pattern": "/^[a-z][a-z0-9_]*$/"
          },
          "name": {
            "type": "string"
          },
          "website": {
            "type": "string",
            "format": "uri"
          },
          "timezone": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "logo_url": {
            "type": "string",
            "format": "uri"
          },
          "settings": {
            "$ref": "#/components/schemas/CreateWorkspaceSettingsDto"
          }
        },
        "required": [
          "id",
          "name",
          "website",
          "timezone",
          "currency"
        ]
      },
      "SmtpSettingsUpdateDto": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "host": {
            "type": "string",
            "maxLength": 255
          },
          "port": {
            "type": "number",
            "minimum": 1,
            "maximum": 65535
          },
          "username": {
            "type": "string",
            "maxLength": 255
          },
          "password_encrypted": {
            "type": "string",
            "maxLength": 255
          },
          "from_name": {
            "type": "string",
            "maxLength": 100
          },
          "from_email": {
            "type": "string",
            "format": "email"
          }
        },
        "required": [
          "enabled",
          "host",
          "port",
          "from_name",
          "from_email"
        ]
      },
      "UpdateWorkspaceSettingsDto": {
        "type": "object",
        "properties": {
          "timescore_reference": {
            "type": "number",
            "minimum": 1
          },
          "bounce_threshold": {
            "type": "number",
            "minimum": 1
          },
          "custom_dimensions": {
            "type": "object"
          },
          "filters": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "integrations": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "geo_enabled": {
            "type": "boolean"
          },
          "geo_store_city": {
            "type": "boolean"
          },
          "geo_store_region": {
            "type": "boolean"
          },
          "geo_coordinates_precision": {
            "type": "number",
            "minimum": 0,
            "maximum": 2
          },
          "annotations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AnnotationDto"
            }
          },
          "smtp": {
            "$ref": "#/components/schemas/SmtpSettingsUpdateDto"
          }
        }
      },
      "UpdateWorkspaceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "initializing",
              "active",
              "inactive",
              "error"
            ]
          },
          "name": {
            "type": "string"
          },
          "website": {
            "type": "string",
            "format": "uri"
          },
          "timezone": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "logo_url": {
            "type": "string",
            "format": "uri"
          },
          "settings": {
            "$ref": "#/components/schemas/UpdateWorkspaceSettingsDto"
          }
        },
        "required": [
          "id"
        ]
      },
...
      "WebsiteMetaDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "format": "uri"
          }
        },
        "required": [
          "url"
        ]
      },
      "WebsiteMetaResponse": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Website title"
          },
          "logo_url": {
            "type": "string",
            "description": "Website logo URL"
          }
        }
      },
...
    }
  }
}